イベントコンポーネントの整理✅
