--- src/pages/eventPage/index.ts
@@ -1,5 +1,0 @@
-export * from "./EventList1";
-export * from "./EventList2";
-export * from "./EventList3";
-export * from "./AllEventList";
-export * from "./AddEventPage";
